src/requirements/local.txt
-r base.txt

black==23.11.0
flake8==6.1.0
isort==5.12.0
psycopg2==2.9.9
watchfiles==0.21.0
locust==2.19.1
pytest-django==4.7.0
pytest-asyncio==0.23.2
factory-boy==3.3.0